+++ users/src/main/java/com/marketplace/users/services/UserService.java
@@ -4,7 +4,7 @@ import com.marketplace.users.models.UserEntity;
 import com.marketplace.users.services.exceptions.InvalidEmailOrPasswordException;
 import com.marketplace.users.services.exceptions.NotFoundEntityException;
 
-public interface UserService extends GenericService<UserEntity, Integer>{
+public interface UserService extends GenericService<UserEntity, Long>{
 
     UserEntity login(String email, String password) throws InvalidEmailOrPasswordException;
